+++ CHANGES.rst
@@ -6,7 +6,8 @@ There's a frood who really knows where his towel is.
 2.3 (unreleased)
 ^^^^^^^^^^^^^^^^
 
-- Nothing changed yet.
+- Added German translations.
+  [tohafi]
 
 
 2.2 (2015-05-04)